audio-kernel: Remove avtimer config
AVTIMER config is enabled to support AV sync for camera recording
usecases. This will result in backward compatibilty issue if the
config is renamed in kernel. To avoid this, have a single point
of control to enable AVTIMER config i.e, in kernel itself.
So, removed it from audio kernel. Update legacy avtimer config
with the config maintained for avtimer feature in kernel.

Change-Id: I90861a0bf1eb33ecc2d173596e70e4a45b0c719c
Signed-off-by: Trinath Thammishetty <tthamish@codeaurora.org>

[Adjusted to use Kconfig]
Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I1aa70356480dea2f0cd76ffada41ccb3d4356b7c