input: fingerprint: correctly set TYPE_FP to /proc/hwinfo
/proc/hwinfo always reported Goodix on devices with either Goodix
or FPC fingerprint sensor. Fix it by updating hwinfo in fpc driver's
prepare stage.

Due to the nature of fingerprint sensor drivers, the update cannot
be performed in the probe stage because both fpc and goodix driver
will be properly probed no matter which sensor the device actually
uses. As a result, the one that is probed later will overwrite the
value in hwinfo.

The fpc blob will even try to enable device_prepare sysfs from
userspace on a device without fpc sensor, and later disable it
again through sysfs. Therefore, we can use this mechanism to toggle
between fpc and goodix in hwinfo. This is acceptable also because
fp_info field in hw_info struct is 1 bit anyway.

Test: boot, cat /proc/hwinfo, and check if the fingerprint value
      matches the userspace logs in logcat when using the
      fingerprint sensor

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: I169ab21b8a41bd801d5039f040754e2ff374a7e0
Signed-off-by: Alex Damaratski <alexeydomoratsky1@gmail.com>