power/supply: bq2591x: Report taper charging status properly
In some conditions, charger board may switch to constant voltage phase for
a few seconds which for whatever reason the driver is reporting it as if it's
not charging, to Qualcomm power stack, causing charging disruptions for a few
seconds during a single charging period.

Add back taper charging reporting originally commented out in OSS release.

Signed-off-by: Albert I <kras@raphielgang.org>