pinctrl: qcom: revert Xiaomi changes
GPIO 0-3, 81-84 and 121 are not accessible on Xiaomi SDM710 based
devices but there are no device configurations in DTS that
use pinctrl config for that gpio pins so there is no reason
to filter them out in msm_config_group_get().
Their filtering in msm_gpio_dbg_show() is also not necessary as
this is handled by previous commits.

Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I37ca5bb1c8d6df83ac3a38365775e9cc84d07694